Remove testbuild warnings
